+++ exchanges_wrapper/exch_srv.py
@@ -12,6 +12,7 @@ import gc
 import traceback
 import asyncio
 import functools
+# noinspection PyPackageRequirements
 import ujson as json
 import logging.handlers
 from decimal import Decimal